+++ src/Set/Defer.php
@@ -274,6 +274,14 @@ final class Defer implements Implementation
     }
 
     /**
+     * @return Implementation<T>
+     */
+    public function memoize(): Implementation
+    {
+        return new Primitive($this->values->memoize());
+    }
+
+    /**
      * @psalm-pure
      */
     private static function distinct(Sequence\Implementation $values): self